Add support for optional query parameters
Closes #2.

For now, we're just adding this for GET requests. So the final
object of a method call will be set as request's `qs` property
for GET requests, and the `body` property for POST and PUT
requests.

An example of using query parameters, implemented in this commit
is as follows:

classy.organizations.retrieveCampaign(34, {
  with: 'organization'
});