Update error handling for non-2XX requests
Before this commit we were handling request errors
that came (typically) from the http.ClientRequest
object by rejecting the native promise set up in
ClassyResource. But we were resolving the promise
for any responses regardless of their statusCode.

With this commit, the http.IncomingMessage response
will be analyzed for 2XX responses. 2XX responses
will resolve the promise and anything else will
reject it.