Add ability to force member token or undefined
Previously we only allowed for the 'app' token to be forced
when making a call to the API. However, there are certain
scenarios where we might want to try a call with a member token
and use 'undefined' if that token doesn't exist, ensuring an
error response from the API.

One example of this scenario is the call to
classy.policyCheck.campaignAdmin(campaignId). We don't want to
make this call as the app because the app always has campaignAdmin
access.